refactor de pantallas login/register y nuevo logo
