+++ app/src/App.css
@@ -1,12 +1,6 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,300;0,400;0,700;0,900;1,100;1,300;1,400;1,700;1,900&display=swap');
-
 .App {
   text-align: center;
   min-height: 100vh;
   display: flex;
   flex-direction: column;
-}
-
-body{
-  font-family: 'Roboto', sans-serif;
 }
