--- app/src/components/addMeme.scss
@@ -1,5 +1,0 @@
-
-.listado-memes{
-    padding:40px;
-    width: 600px;
-}
