--- app/src/components/comment.scss
@@ -1,1 +1,0 @@
-
