--- app/src/components/header.scss
@@ -1,59 +1,0 @@
-.navbar-brand{
-	background-image: url('../assets/images/logo.png');
-	width: 53px;
-	height: 54px;
-	background-size: 100%;
-	background-repeat: no-repeat;
-	padding: 10px;
-}
-
-.bg-top {
-	background-color:#3e3e3e;
-height: 65px;
-z-index: 9999;
-}
-
-.menu{
-	position: absolute;
-	right: 20px;
-  }
-
-.nav-item a{
-	color:white !important;
-	margin-left:10px;
-  }
-
-
-li {
-  list-style-type: none;
-}
-
-.dropleft .dropdown-menu {
-	width: 30vw;
-	height: 100vh;
-    top: -14px;
-    right: -23px;
-    left: auto;
-    margin-top: 0;
-    margin-right: .125rem;
-	border-radius: 0px;
-	background:#E33071;
-	padding-top: 65px;
-	li{
-		list-style-type: none;
-		text-align: center;
-		padding: 20px;
-		border-bottom: 2px solid #0000000a;
-		a{
-			color:white !important;
-		}
-	}
-	li:hover{
-		background:white;
-		a{
-		color:#E1BD3B !important;
-		text-decoration: none;
-		font-weight: bold;
-		}
-	}
-}
