--- app/src/components/listCategory.scss
@@ -1,10 +1,0 @@
-.listado-categorias {
-	padding: 40px;
-	a {
-		font-size: 14px;
-	}
-}
-.list-group-item-action:focus {
-	color: white;
-	background-color: #3e3e3e;
-}
