--- app/src/components/listMeme.scss
@@ -1,4 +1,0 @@
-.listado-memes{
-    padding:40px;
-    width: 600px;
-}
