--- app/src/components/login.scss
@@ -1,115 +1,0 @@
-.login-wrapper{
-    padding:80px;
-}
-
-@import url(https://fonts.googleapis.com/css?family=Roboto:300);
-
-#img1{
-    position: relative;
-    top: 23%;
-    right: 11%;
-    width: 275px !important;
-    height: 206px !important;
-}
-
-#img2{
-    position: relative;
-    top: -4%;
-    right: -14%;
-    width: 300px !important;
-    height: 196px !important;
-}
-
-.form {
-  position: relative;
-  z-index: 1;
-  background: #FFFFFF;
-  max-width: 360px;
-  padding: 0;
-  text-align: center;
-}
-.form input {
-  font-family: "Roboto", sans-serif;
-  outline: 0;
-  background: #f2f2f2;
-  width: 100%;
-  border: 0;
-  margin: 0 0 15px;
-  padding: 15px;
-  box-sizing: border-box;
-  font-size: 14px;
-}
-.submit {
-  background: #E33071 !important;
-}
-
-.form a {
-    font-family: "Roboto", sans-serif;
-    text-transform: uppercase;
-    outline: 0;
-    background: #E1BD3B;
-    width: 100%;
-    border: 0;
-    padding: 15px;
-    color: #FFFFFF;
-    font-size: 14px;
-    -webkit-transition: all 0.3 ease;
-    transition: all 0.3 ease;
-    cursor: pointer; 
-    display: block;
-    text-decoration: none;
-}
-
-.submit:hover,.submit:active,.submit:focus {
-  background: black !important;
-}
-
-.form a:hover,.form a:active,.form a:focus {
-    background: black ;
-}
-
-.form .message {
-  margin: 15px 0 0;
-  color: #b3b3b3;
-  font-size: 12px;
-}
-.form .message a {
-  color: #4CAF50;
-  text-decoration: none;
-}
-.form .register-form {
-  display: none;
-}
-.container {
-  position: relative;
-  z-index: 1;
-  max-width: 300px;
-  margin: 0 auto;
-}
-.container:before, .container:after {
-  content: "";
-  display: block;
-  clear: both;
-}
-.container .info {
-  margin: 50px auto;
-  text-align: center;
-}
-.container .info h1 {
-  margin: 0 0 15px;
-  padding: 0;
-  font-size: 36px;
-  font-weight: 300;
-  color: #1a1a1a;
-}
-.container .info span {
-  color: #4d4d4d;
-  font-size: 12px;
-}
-.container .info span a {
-  color: #000000;
-  text-decoration: none;
-}
-.container .info span .fa {
-  color: #EF3B3A;
-}
