--- app/src/components/main.scss
@@ -1,30 +1,0 @@
-.borderless {
-	border-top: 0 none;
-}
-
-.btn-vote-up {
-	height: 40px;
-	background: #6bddbb;
-	border-radius: 20px;
-}
-
-.btn-vote-down {
-	height: 40px;
-	background: #000;
-	color: white;
-	border-radius: 20px;
-}
-.btn-vote-down:hover {
-	color: white;
-}
-.circuloComment {
-	width: 38px;
-	height: 38px;
-	background: #e1bd3b;
-}
-
-.circuloAddComment {
-	width: 51px;
-	height: 51px;
-	background: #e33071;
-}
