--- app/src/components/memePreview.scss
@@ -1,30 +1,0 @@
-.titulo-meme{
-	font-weight: bold;
-	color:#4F4F4F;
-	small{
-		color:#E1BD3B;
-		font-size:14px;
-		font-weight: 300;
-	}
-}
-
-.small-link{
-	text-decoration: none;
-	background-color:
-	transparent;
-	font-size: 13px;
-	color:black;
-	font-weight: 300;
-}
-.small-link:hover{
-	text-decoration: none;
-	background-color:
-	transparent;
-	font-size: 13px;
-	color:#e33071;
-	font-weight: 300;
-}
-
-.preview-wrapper{
-	margin-bottom:50px;
-}
